refact : Dealer, Player 상위타입 추상 클래스 추가
