+++ src/main/kotlin/blackjack/domain/gamer/BlackJackGamer.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gamer
 
 import blackjack.domain.card.Card
