+++ src/main/kotlin/blackjack/domain/gamer/GameRecordType.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gamer
 
 enum class GameRecordType(val value: String) {
     WIN("승"), LOSE("패"), DRAW("무"), NONE("NONE")
