+++ src/main/kotlin/blackjack/domain/gamer/GamerType.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gamer
 
 enum class GamerType {
     PLAYER, DEALER
