+++ src/test/kotlin/blackjack/domain/RuleCheckerTest.kt
@@ -3,6 +3,9 @@ package blackjack.domain
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardNumber
 import blackjack.domain.card.CardShape
+import blackjack.domain.gamer.Dealer
+import blackjack.domain.gamer.GameRecordType
+import blackjack.domain.gamer.Player
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
