+++ src/test/kotlin/blackjack/domain/gamer/PlayerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gamer
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardNumber
